Refactor code

+++ src/main/java/com/qa/ims/controller/OrderController.java
@@ -16,9 +16,9 @@ import com.qa.ims.persistence.domain.OrderLineItem;
 import com.qa.ims.utils.Utils;
 
 public class OrderController implements CrudController<Order> {
-	
+
 	public static final Logger LOGGER = LogManager.getLogger();
-	
+
 	private OrderDAO orderDAO;
 	private Utils utils;
 
@@ -27,15 +27,28 @@ public class OrderController implements CrudController<Order> {
 		this.utils = utils;
 	}
 
+	/**
+	 * Reads all orders to the logger
+	 */
 	@Override
 	public List<Order> readAll() {
 		List<Order> orders = orderDAO.readAll();
-		for (Order order : orders) {
-			LOGGER.info(order);
+		if (orders.size() > 0) {
+			for (Order order : orders) {
+				LOGGER.info(order);
+			}
+		} else {
+			LOGGER.info("There are currently no orders in the system");
 		}
+
 		return orders;
 	}
 
+	/**
+	 * Creates an order by taking in user input
+	 * Only attempts to create the order if a valid
+	 * customer id has been input.
+	 */
 	@Override
 	public Order create() {
 		CustomerDAO custDAO = new CustomerDAO();
@@ -44,11 +57,22 @@ public class OrderController implements CrudController<Order> {
 		LOGGER.info("Please enter a customer id");
 		Long customerId = utils.getLong();
 		Customer customer = custDAO.read(customerId);
-		Order order = orderDAO.create(new Order(orderNumber, customer));
-		LOGGER.info("Order created");
-		return order;
+		if (customer != null) {
+			Order order = orderDAO.create(new Order(orderNumber, customer));
+			LOGGER.info("Order created");
+			return order;	
+		} else {
+			LOGGER.info("Please try again with a valid customer");
+			return null;
+		}
 	}
 
+	
+	/**
+	 * Updates an existing order by taking in user input
+	 * Only attempts to update the order if a valid
+	 * customer id has been input.
+	 */
 	@Override
 	public Order update() {
 		CustomerDAO custDAO = new CustomerDAO();
@@ -59,60 +83,100 @@ public class OrderController implements CrudController<Order> {
 		LOGGER.info("Please enter a customer id");
 		Long customerId = utils.getLong();
 		Customer customer = custDAO.read(customerId);
-		Order order = orderDAO.update(new Order(id, orderNumber, customer));
-		LOGGER.info("Order updated");
-		return order;
+		if (customer != null) {
+			Order order = orderDAO.update(new Order(id, orderNumber, customer));
+			if (order != null) {
+				LOGGER.info("Order updated");
+			}
+			return order;
+		}
+
+		return null;
+
 	}
+
 	
+	/**
+	 * Adds an item to an order.
+	 * Only attempts to create the line item if valid order and customer
+	 * ids have been added. Calls the OrderLineItemController method for
+	 * the database updates. 
+	 */
 	public Order addItem() {
 		ItemDAO itemDAO = new ItemDAO();
-		OrderLineItemDAO lineItemDAO = new OrderLineItemDAO();
+		OrderLineItemController lineItemController = new OrderLineItemController();
+
 		LOGGER.info("Please enter the order id");
 		Long orderId = utils.getLong();
 		LOGGER.info("Please enter the item id");
 		Long itemId = utils.getLong();
-		Item item = itemDAO.read(itemId);
 		LOGGER.info("Please enter the item quantity");
 		Long quantity = utils.getLong();
-		OrderLineItem currentItem = lineItemDAO.readByOrderItem(orderId, itemId);
-		OrderLineItem lineItem;
-		if (currentItem != null) {
-			currentItem.setQuantity(currentItem.getQuantity() + quantity);
-			lineItem = lineItemDAO.update(currentItem);
+		Item item = itemDAO.read(itemId);
+		Order order = orderDAO.read(orderId);
+		if (item != null && order != null) {
+			lineItemController.addToOrder(item, orderId, itemId, quantity);
+			Order updatedOrder = calculateTotal(orderId);
+			LOGGER.info(quantity + " of " + item.getName() + " added");
+			return updatedOrder;
 		} else {
-			lineItem = lineItemDAO.create(new OrderLineItem(item, quantity, orderId));
+			LOGGER.info("Some information you entered is incorrect, please try again");
+			return null;
 		}
-		item.updateStock(-quantity);
-		itemDAO.update(item);
-		Order order = calculateTotal(orderId);
-		LOGGER.info(quantity + " of " + lineItem.getItem().getName() + " added");
-		return order;
+
 	}
-	
+
+	/**
+	 * Removes an item from an order.
+	 * Only attempts to delete the line item if valid order and customer
+	 * ids have been added. 
+	 */
 	public Order removeItem() {
 		OrderLineItemDAO lineItemDAO = new OrderLineItemDAO();
 		ItemDAO itemDAO = new ItemDAO();
+
 		LOGGER.info("Please enter the order id");
 		Long orderId = utils.getLong();
 		LOGGER.info("Please enter the id of the item to remove");
 		Long itemId = utils.getLong();
 		Item item = itemDAO.read(itemId);
-		OrderLineItem lineItem = lineItemDAO.readByOrderItem(orderId, itemId);
-		lineItemDAO.delete(lineItem.getId());
-		item.updateStock(lineItem.getQuantity());
-		itemDAO.update(item);
-		Order order = calculateTotal(orderId);
-		return order;
+		Order order = orderDAO.read(orderId);
+		if (item == null || order == null) {
+			LOGGER.info("Some information you entered is incorrect, please try again");
+			return null;
+		} else {
+			OrderLineItem lineItem = lineItemDAO.readByOrderItem(orderId, itemId);
+			if (lineItem != null) {
+				lineItemDAO.delete(lineItem.getId());
+				item.updateStock(lineItem.getQuantity());
+				itemDAO.update(item);
+				Order updatedOrder = calculateTotal(orderId);
+				return updatedOrder;
+			} else {
+				LOGGER.info("This order does not contain the selected item.");
+				return null;
+			}
+		}
+
 	}
-	
+
+	/**
+	 * Updates the order total value in the database based on
+	 * the line items.
+	 * The Order class instance total is updated automatically when the order
+	 * is read.
+	 */
 	public Order calculateTotal(Long orderId) {
 		Order selectedOrder = orderDAO.read(orderId);
 		Order revisedOrder = orderDAO.update(selectedOrder);
-		
+
 		return revisedOrder;
-		
+
 	}
 
+	/**
+	 * Deletes an existing order by taking in user input
+	 */
 	@Override
 	public int delete() {
 		LOGGER.info("Please enter the id of the order you would like to delete");
